requirements_professional.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Web Requests and APIs
requests>=2.28.0
python-dotenv>=1.0.0
yfinance>=0.2.0

# Visualization and Dashboard
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.14.0
dash-table>=5.0.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0

# Database and Storage
sqlalchemy>=2.0.0
sqlite3

# Mathematical and Statistical Analysis
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Additional Professional Tools
openpyxl>=3.1.0  # Excel export
tabulate>=0.9.0  # Professional table formatting
colorama>=0.4.6  # Colored terminal output
tqdm>=4.65.0     # Progress bars
click>=8.1.0     # Better CLI interface

# Performance and Optimization
numba>=0.57.0    # JIT compilation for speed
joblib>=1.3.0    # Parallel processing

# Data Validation
pydantic>=2.0.0

# Logging and Monitoring
psutil>=5.9.0    # System monitoring
python-json-logger>=2.0.0

# Development and Testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0